+++ packages/core/src/components/ozon-content/ozon-content-mapper.tsx
@@ -18,6 +18,7 @@ import { getNodeName } from './get-node-name.function';
 import { OzonContentNode } from './ozon-content-node.interface';
 import { OzonContentContext } from './ozon-content-context.interface';
 import { OzonContentFallbackNode } from './nodes/fallback.node';
+import { OzonContentLijstNode } from './nodes/lijst.node';
 
 export class Mapper {
   private cache: { xml: string, document: Document } | undefined;
@@ -36,6 +37,7 @@ export class Mapper {
     new OzonContentTableNode(),
     new OzonContentIntIoRefNode(),
     new OzonContentFiguurNode(),
+    new OzonContentLijstNode(),
   ];
 
   private skip = this.mappers.reduce<string[]>((t, m) => {