+++ packages/sources/src/components/ozon-content/ozon-content.content.ts
@@ -14,6 +14,119 @@ export const content = [
     `,
   },
   {
+    title: 'Inhoud met Lijst',
+    content: `
+      <Inhoud>
+        <Lijst
+          eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1"
+          type="expliciet"
+          wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1"
+        >
+          <Lijstaanhef>Deze afdeling is niet van toepassing op:</Lijstaanhef>
+          <Li
+            eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_a"
+            wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_a"
+          >
+            <LiNummer>a.</LiNummer>
+            <Al>wonen;</Al>
+          </Li>
+          <Li
+            eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_b"
+            wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_b"
+          >
+            <LiNummer>b.</LiNummer>
+            <Al>
+              het feitelijk verrichten van bouw- en sloopwerkzaamheden aan bouwwerken of
+              het feitelijk verrichten van onderhoudswerkzaamheden aan een bouwwerk of
+              van een terrein;
+            </Al>
+          </Li>
+          <Li
+            eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_c"
+            wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_c"
+          >
+            <LiNummer>c.</LiNummer>
+            <Al>
+              een milieubelastende activiteit die in hoofdzaak in de openbare
+              buitenruimte wordt verricht;
+            </Al>
+          </Li>
+        </Lijst>
+
+        <Al>Alde Feanen behoort tot het Natura 2000-landschap 'Meren en Moerassen'.</Al>
+        <Al>De begrenzingsmethodiek is verder uitgewerkt in het Gebiedendocument (2004).</Al>
+
+        <Lijst
+          eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1"
+          type="expliciet"
+          wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1"
+        >
+          <Li
+            eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_d"
+            wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_d"
+          >
+            <LiNummer>d.</LiNummer>
+            <Al>doorgaand verkeer op wegen, vaarwegen en spoorwegen;</Al>
+          </Li>
+          <Li
+            eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e"
+            wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e"
+          >
+            <LiNummer>e.</LiNummer>
+            <Lijst
+              eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1"
+              type="expliciet"
+              wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1"
+            >
+              <Lijstaanhef>een evenement:</Lijstaanhef>
+              <Li
+                eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1__item_1"
+                wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1__item_1"
+              >
+                <LiNummer>1°.</LiNummer>
+                <Al>
+                  dat ergens anders plaatsvindt dan op een locatie voor evenementen;
+                </Al>
+              </Li>
+              <Li
+                eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1__item_2"
+                wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1__item_2"
+              >
+                <LiNummer>2°.</LiNummer>
+                <Al>
+                  dat geen festiviteit als bedoeld in artikel 5.68 van het Besluit
+                  kwaliteit leefomgeving is; of
+                </Al>
+              </Li>
+              <Li
+                eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1__item_3"
+                wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_e__list_o_1__item_3"
+              >
+                <LiNummer>3°.</LiNummer>
+                <Al>
+                  waarover geluidregels zijn gesteld bij of krachtens een gemeentelijke
+                  verordening; en
+                </Al>
+              </Li>
+            </Lijst>
+          </Li>
+          <Li
+            eId="chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_f"
+            wId="gm1979_2__chp_3__subchp_3.4__subsec_3.4.1__art_3.41__para_2__list_o_1__item_f"
+          >
+            <LiNummer>f.</LiNummer>
+            <Al>
+              het verrichten van werkzaamheden met een mobiele installatie op een
+              weiland, akker of bos die geen verplaatsbaar mijnbouwwerk als bedoeld in
+              artikel 4.1116 van het Besluit activiteiten leefomgeving is.
+            </Al>
+          </Li>
+          <Lijstsluiting>Dit is de plaats om een afsluitende zin onder de lijst te plaatsen.</Lijstsluiting>
+        </Lijst>
+      </Inhoud>
+    `,
+  },
+  {
     // De Noot in deze content is verzonnen. Volgens STOP kan dit voorkomen maar ik heb geen echte content gevonden.
     title: 'Opschrift',
     content: `
@@ -140,6 +253,7 @@ export const content = [
             artikel 4.1116 van het Besluit activiteiten leefomgeving is.
           </Al>
         </Li>
+        <Lijstsluiting>Dit is de plaats om een afsluitende zin onder de lijst te plaatsen.</Lijstsluiting>
       </Lijst>
     `
   },